+++ hoomd/md/HarmonicImproperForceComputeGPU.cc
@@ -61,7 +61,7 @@ void HarmonicImproperForceComputeGPU::setParams(unsigned int type, Scalar K, Sca
 
     Calls gpu_compute_harmonic_improper_forces to do the dirty work.
 */
-void HarmonicImproperForceComputeGPU::computeForces(unsigned int timestep)
+void HarmonicImproperForceComputeGPU::computeForces(uint64_t timestep)
     {
     // start the profile
     if (m_prof) m_prof->push(m_exec_conf, "Harmonic Improper");
